+++ rtx_core/src/definition/conditional.rs
@@ -167,13 +167,13 @@ impl Conditional {
 
         // print STDERR "{true}\n" if $tracing;
       } else {
-        let _to = self.skip_conditional_body(-1, gullet, state);
+        let _to = self.skip_conditional_body(-1.0, gullet, state);
         // print STDERR "{false} [skipped to " . ToString($to) . "]\n" if $tracing;
       }
     } else {
       // If there's no test, it must be the Special Case, \ifcase
       let num = args[0].to_number().value_of();
-      if num > 0 {
+      if num > 0.0 {
         let _to = self.skip_conditional_body(num, gullet, state);
         //       print STDERR "{$num} [skipped to " . ToString($to) . "]\n" if $tracing;
       }
@@ -205,9 +205,9 @@ impl Conditional {
   // #   \if\ifx AA XY junk \else blah \fi True \else False \fi
   // # The inner \ifx should expand to "XY junk", since A==A
   // # Return the token we've skipped to, and the frame that this applies to.
-  fn skip_conditional_body(&self, nskips: i32, gullet: &mut Gullet, state: &mut State) -> Tokens {
+  fn skip_conditional_body(&self, nskips: f32, gullet: &mut Gullet, state: &mut State) -> Tokens {
     let mut level = 1;
-    let mut n_ors = 0;
+    let mut n_ors = 0.0;
     let _start = gullet.get_locator();
     // NOTE: Open-coded manipulation of if_stack!
     // [we're only reading tokens & looking up, so State shouldn't change behind our backs]
@@ -246,11 +246,11 @@ impl Conditional {
           if level <= 1 {
             // Ignore \else,\or nested in the body.
             if cond_type == Some(ConditionalType::Or) {
-              n_ors += 1;
+              n_ors += 1.0;
               if n_ors == nskips {
                 return t;
               }
-            } else if cond_type == Some(ConditionalType::Else) && nskips != 0 {
+            } else if cond_type == Some(ConditionalType::Else) && nskips != 0.0 {
               // Found \else and we're looking for one?
               // Make sure this \else is NOT for a nested \if that is part of the test clause!
               if let Some(Stored::VecDequeStored(stack)) = state.lookup_value_mut("if_stack") {
@@ -303,7 +303,7 @@ impl Conditional {
         Ok(Tokens!())
       } else {
         state.if_frame = Some(stack_frame.clone());
-        let t = self.skip_conditional_body(0, gullet, state);
+        let t = self.skip_conditional_body(0.0, gullet, state);
         //     print STDERR '{' . ToString($LaTeXML::CURRENT_TOKEN) . '}'
         //       . " [for " . ToString($$LaTeXML::IFFRAME{token}) . " #" .
         // $$LaTeXML::IFFRAME{ifid}       . " skipping to " . ToString($t) . "]\n"
